Wip reorganize BG3 utils files
