+++ src/Games/NexusMods.Games.Larian/BaldursGate3/Diagnostics.cs
@@ -18,16 +18,24 @@ internal static partial class Diagnostics
         .WithId(new DiagnosticId(Source, number: 1))
         .WithTitle("Missing required dependency")
         .WithSeverity(DiagnosticSeverity.Warning)
-        .WithSummary("Mod {PakMod} is missing required dependency '{MissingDependencyName}'.")
+        .WithSummary("The mod `{ModName}` is missing the required dependency '{MissingDepName}' v{MissingDepVersion}.")
         .WithDetails("""
-                     '{MissingDependencyName}' is required by '{PakModuleName}' but is not present in the loadout.
+                     
+                     '{MissingDependencyName}' v{MissingDevVersion} is not installed or enabled in your Loadout. This pak module is required by `{PakModuleName}` v{PakModuleVersion} to run correct.
+                     
+                     ## Recommended actions
+                     ### Search for and install the missing mod
+                     You can search for '{MissingDependencyName}' on {NexusModsLink}
+                     
                      
                      You can try to search the missing mod on {NexusModsLink} or using the in-game mod manager.
                      """)
         .WithMessageData(messageBuilder => messageBuilder
-            .AddDataReference<LoadoutItemGroupReference>("PakMod")
-            .AddValue<string>("MissingDependencyName")
+            .AddDataReference<LoadoutItemGroupReference>("ModName")
+            .AddValue<string>("MissingDepName")
             .AddValue<string>("PakModuleName")
+            .AddValue<string>("PakModuleVersion")
+            .AddValue<string>("MissingDepVersion")
             .AddValue<NamedLink>("NexusModsLink")
         )
         .Finish();
