+++ src/Games/NexusMods.Games.Larian/BaldursGate3/Pipelines.cs
@@ -6,6 +6,7 @@ using NexusMods.Abstractions.IO;
 using NexusMods.Abstractions.Resources;
 using NexusMods.Abstractions.Resources.Caching;
 using NexusMods.Abstractions.Resources.IO;
+using NexusMods.Games.Larian.BaldursGate3.Utils.LsxXmlParsing;
 using NexusMods.Games.Larian.BaldursGate3.Utils.PakParsing;
 using NexusMods.Hashing.xxHash64;
