+++ src/Games/NexusMods.Games.Larian/BaldursGate3/Utils/LsxXmlParsing/ModsettingsFileFormat.cs
@@ -1,7 +1,6 @@
-using System.Text;
 using System.Xml;
 
-namespace NexusMods.Games.Larian.BaldursGate3.Utils.PakParsing;
+namespace NexusMods.Games.Larian.BaldursGate3.Utils.LsxXmlParsing;
 
 /// <summary>
 /// Class to parse and write the `modsettings.lsx` xml file format used for the BG3 Load Order.
